edit request, delete modal

+++ src/app/components/tables/transaction-table/transaction-table.component.ts
@@ -2,6 +2,10 @@ import { HttpClient } from "@angular/common/http";
 import { Component, OnInit } from "@angular/core";
 import { TransactionServiceService } from "src/app/shared/services/transaction-service.service";
 import { TransactionUpdateService } from "src/app/shared/services/transaction-update.service";
+import { DeleteModalComponent } from "../../modal/delete-modal/delete-modal.component";
+import { MatDialog } from "@angular/material/dialog";
+import { MatSnackBar } from "@angular/material/snack-bar";
+import { EditModalComponent } from "../../modal/edit-modal/edit-modal.component";
 
 @Component({
   selector: "app-transaction-table",
@@ -14,7 +18,9 @@ export class TransactionTableComponent implements OnInit {
   constructor(
     private transactionService: TransactionServiceService,
     private transactionUpdateService: TransactionUpdateService,
-    private http: HttpClient
+    private http: HttpClient,
+    public dialog: MatDialog,
+    public snackBar: MatSnackBar
   ) {}
 
   userID: string = "";
@@ -29,6 +35,16 @@ export class TransactionTableComponent implements OnInit {
     this.transactionUpdateService.transactionAdded.subscribe((transaction) => {
       this.addTransactionToList(transaction);
     });
+    this.transactionUpdateService.transactionUpdated.subscribe(
+      (updatedTransaction) => {
+        const index = this.transactionList.findIndex(
+          (transaction) => transaction._id === updatedTransaction._id
+        );
+        if (index !== -1) {
+          this.transactionList[index] = updatedTransaction;
+        }
+      }
+    );
   }
 
   addTransactionToList(transaction: any) {
@@ -47,17 +63,45 @@ export class TransactionTableComponent implements OnInit {
     );
   }
 
+  openSnackBar(message: string, action: string) {
+    this.snackBar.open(message, action, {
+      duration: 5000,
+    });
+  }
+
+  updateTransaction(transaction: any) {
+    const dialogRef = this.dialog.open(EditModalComponent, {
+      width: "370px",
+      data: transaction,
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      console.log("The dialog was closed");
+    });
+  }
+
   deleteTransaction(transactionId: string) {
-    console.log(transactionId);
-    const url = `http://localhost:3000/delete/transactions/${transactionId}`;
-    this.http.delete(url).subscribe(
-      (response: any) => {
-        console.log("Transaction deleted successfully");
-        this.getTransactions(this.userID);
+    const dialogRef = this.dialog.open(DeleteModalComponent, {
+      data: {
+        message: "Do you want to delete?",
       },
-      (error) => {
-        console.error("Error:", error);
+    });
+
+    dialogRef.afterClosed().subscribe((confirmed: boolean) => {
+      if (confirmed) {
+        const url = `http://localhost:3000/delete/transactions/${transactionId}`;
+        this.http.delete(url).subscribe(
+          (response: any) => {
+            console.log("Transaction deleted successfully");
+            this.getTransactions(this.userID);
+          },
+          (error) => {
+            console.error("Error:", error);
+          }
+        );
+      } else {
+        this.openSnackBar("canceled deleting", "ok");
       }
-    );
+    });
   }
 }
